requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
tzdata==2024.2
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.2
tzdata==2024.2
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Markdown==3.7
sqlparse==0.5.2
tzdata==2024.2
